kernelci.test: derive compression type from URLs
Rather than hard-coding the compression type in the LAVA base
template, derive it from the rootfs and modules URLs and pass it as
template arguments.  This allows more flexibility by relying on the
convention with usual file extensions i.e. .gz, .bz2 and .xz.  If the
end of the URL doesn't match any of these (say, .tar), the compression
attribute is left empty and not defined in the LAVA template.

Signed-off-by: Guillaume Tucker <guillaume.tucker@collabora.com>
Signed-off-by: Alice Ferrazzi <alice.ferrazzi@miraclelinux.com>